test(LottoTest): 로또 티켓 단위 테스트
- 기존 LottoSheet 삭제에 따라 테스트 코드도 삭제
